+++ packages/expo-router/src/global-state/routing.ts
@@ -28,29 +28,24 @@ import { Href } from '../types';
 import { SingularOptions } from '../useScreens';
 import { shouldLinkExternally } from '../utils/url';
 
-function isCurrentNavigatorReady() {
-  let state = store.navigationRef.getRootState();
-  while (state) {
-    const index = state.index ?? 0;
-    const route = state.routes[index];
-    if (route.params && 'screen' in route.params && !route.state) {
-      return false;
-    }
-    state = route.state;
-  }
-  return true;
-}
-
 function assertIsReady() {
-  if (!store.navigationRef.isReady() || !isCurrentNavigatorReady()) {
+  if (!store.navigationRef.isReady()) {
     throw new Error(
       'Attempted to navigate before mounting the Root Layout component. Ensure the Root Layout component is rendering a Slot, or other navigator on the first render.'
     );
   }
 }
 
+interface LinkAction {
+  type: 'ROUTER_LINK';
+  payload: {
+    options: LinkToOptions;
+    href: string;
+  };
+}
+
 export const routingQueue = {
-  queue: [] as NavigationAction[],
+  queue: [] as (NavigationAction | LinkAction)[],
   subscribers: new Set<() => void>(),
   subscribe(callback: () => void) {
     routingQueue.subscribers.add(callback);
@@ -61,7 +56,7 @@ export const routingQueue = {
   snapshot() {
     return routingQueue.queue;
   },
-  add(action: NavigationAction) {
+  add(action: NavigationAction | LinkAction) {
     routingQueue.queue.push(action);
     for (const callback of routingQueue.subscribers) {
       callback();
@@ -71,10 +66,28 @@ export const routingQueue = {
     // Reset the identity of the queue.
     const events = routingQueue.queue;
     routingQueue.queue = [];
-    let action: NavigationAction | undefined;
+    let action: NavigationAction | LinkAction | undefined;
     while ((action = events.shift())) {
       if (ref.current) {
-        ref.current.dispatch(action);
+        if (action.type === 'ROUTER_LINK') {
+          const {
+            payload: { href, options },
+          } = action as LinkAction;
+
+          action = getNavigateAction(
+            href,
+            options,
+            options.event,
+            options.withAnchor,
+            options.dangerouslySingular,
+            !!options.__internal__PreviewKey
+          );
+          if (action) {
+            ref.current.dispatch(action);
+          }
+        } else {
+          ref.current.dispatch(action);
+        }
       }
     }
   },
@@ -219,6 +232,44 @@ export function linkTo(originalHref: Href, options: LinkToOptions = {}) {
     return;
   }
 
+  if (href === '..' || href === '../') {
+    assertIsReady();
+    const navigationRef = store.navigationRef.current;
+
+    if (navigationRef == null) {
+      throw new Error(
+        "Couldn't find a navigation object. Is your component inside NavigationContainer?"
+      );
+    }
+
+    if (!store.linking) {
+      throw new Error('Attempted to link to route when no routes are present');
+    }
+
+    navigationRef.goBack();
+    return;
+  }
+
+  const linkAction: LinkAction = {
+    type: 'ROUTER_LINK',
+    payload: {
+      href,
+      options,
+    },
+  };
+
+  routingQueue.add(linkAction);
+}
+
+function getNavigateAction(
+  baseHref: string,
+  options: LinkToOptions,
+  type = 'NAVIGATE',
+  withAnchor?: boolean,
+  singular?: SingularOptions,
+  isPreviewNavigation?: boolean
+) {
+  let href: string | undefined = baseHref;
   assertIsReady();
   const navigationRef = store.navigationRef.current;
 
@@ -227,20 +278,13 @@ export function linkTo(originalHref: Href, options: LinkToOptions = {}) {
       "Couldn't find a navigation object. Is your component inside NavigationContainer?"
     );
   }
-
   if (!store.linking) {
     throw new Error('Attempted to link to route when no routes are present');
   }
-
-  if (href === '..' || href === '../') {
-    navigationRef.goBack();
-    return;
-  }
-
   const rootState = navigationRef.getRootState();
 
   href = resolveHrefStringWithSegments(href, store.getRouteInfo(), options);
-  href = applyRedirects(href, store.redirects);
+  href = applyRedirects(href, store.redirects) ?? undefined;
 
   // If the href is undefined, it means that the redirect has already been handled the navigation
   if (!href) {
@@ -253,27 +297,6 @@ export function linkTo(originalHref: Href, options: LinkToOptions = {}) {
     console.error('Could not generate a valid navigation state for the given path: ' + href);
     return;
   }
-
-  routingQueue.add(
-    getNavigateAction(
-      state,
-      rootState,
-      options.event,
-      options.withAnchor,
-      options.dangerouslySingular,
-      !!options.__internal__PreviewKey
-    )
-  );
-}
-
-function getNavigateAction(
-  _actionState: ResultState,
-  _navigationState: NavigationState,
-  type = 'NAVIGATE',
-  withAnchor?: boolean,
-  singular?: SingularOptions,
-  isPreviewNavigation?: boolean
-) {
   /**
    * We need to find the deepest navigator where the action and current state diverge, If they do not diverge, the
    * lowest navigator is the target.
@@ -290,8 +313,8 @@ function getNavigateAction(
    */
 
   const { actionStateRoute, navigationState } = findDivergentState(
-    _actionState,
-    _navigationState,
+    state,
+    rootState,
     type === 'PRELOAD'
   );